US #718053. [Jenkins] update jenkins plugin to work with latest SDK version 2.x
