+++ src/main/java/com/microfocus/application/automation/tools/octane/events/SCMListenerOctaneImpl.java
@@ -21,17 +21,16 @@
 package com.microfocus.application.automation.tools.octane.events;
 
 import com.hp.octane.integrations.OctaneSDK;
-import com.microfocus.application.automation.tools.octane.configuration.ConfigurationService;
-import com.microfocus.application.automation.tools.octane.model.processors.scm.SCMProcessor;
-import com.microfocus.application.automation.tools.octane.tests.build.BuildHandlerUtils;
 import com.hp.octane.integrations.dto.DTOFactory;
 import com.hp.octane.integrations.dto.events.CIEvent;
 import com.hp.octane.integrations.dto.events.CIEventType;
 import com.hp.octane.integrations.dto.scm.SCMData;
+import com.microfocus.application.automation.tools.octane.configuration.ConfigurationService;
 import com.microfocus.application.automation.tools.octane.model.CIEventCausesFactory;
+import com.microfocus.application.automation.tools.octane.model.processors.scm.SCMProcessor;
 import com.microfocus.application.automation.tools.octane.model.processors.scm.SCMProcessors;
+import com.microfocus.application.automation.tools.octane.tests.build.BuildHandlerUtils;
 import hudson.Extension;
-import hudson.FilePath;
 import hudson.matrix.MatrixConfiguration;
 import hudson.model.AbstractBuild;
 import hudson.model.Run;
@@ -39,12 +38,9 @@ import hudson.model.TaskListener;
 import hudson.model.listeners.SCMListener;
 import hudson.scm.ChangeLogSet;
 import hudson.scm.SCM;
-import hudson.scm.SCMRevisionState;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
-
-import java.io.File;
 
 /**
  * Run Listener that handles SCM CI events and dispatches notifications to the Octane server
@@ -77,7 +73,7 @@ public class SCMListenerOctaneImpl extends SCMListener {
 			SCMData scmData = extractSCMData(run, scm, scmProcessor);
 			if (scmData != null) {
 				CIEvent event = createSCMEvent(run, scmData);
-				OctaneSDK.getInstance().getEventsService().publishEvent(event);
+				OctaneSDK.getClients().forEach(client ->client.getEventsService().publishEvent(event));
 			}
 		} catch (Throwable throwable) {
 			logger.error("failed to build and/or dispatch SCM event for " + run, throwable);
