+++ src/main/java/com/microfocus/application/automation/tools/octane/events/WorkflowListenerOctaneImpl.java
@@ -23,7 +23,6 @@ package com.microfocus.application.automation.tools.octane.events;
 import com.google.inject.Inject;
 import com.hp.octane.integrations.OctaneSDK;
 import com.hp.octane.integrations.dto.DTOFactory;
-
 import com.hp.octane.integrations.dto.events.CIEvent;
 import com.hp.octane.integrations.dto.events.CIEventType;
 import com.hp.octane.integrations.dto.events.MultiBranchType;
@@ -100,7 +99,7 @@ public class WorkflowListenerOctaneImpl implements GraphListener {
 					.setProjectDisplayName(parentRun.getParent().getFullName());
 		}
 
-		OctaneSDK.getInstance().getEventsService().publishEvent(event);
+		OctaneSDK.getClients().forEach(octaneClient -> octaneClient.getEventsService().publishEvent(event));
 	}
 
 	private void sendPipelineFinishedEvent(FlowEndNode flowEndNode) {
@@ -119,7 +118,7 @@ public class WorkflowListenerOctaneImpl implements GraphListener {
 				.setResult(BuildHandlerUtils.translateRunResult(parentRun))
 				.setCauses(CIEventCausesFactory.processCauses(parentRun))
 				.setTestResultExpected(hasTests);
-		OctaneSDK.getInstance().getEventsService().publishEvent(event);
+		OctaneSDK.getClients().forEach(octaneClient -> octaneClient.getEventsService().publishEvent(event));
 	}
 
 	private void sendStageStartedEvent(StepStartNode stepStartNode) {
@@ -134,7 +133,7 @@ public class WorkflowListenerOctaneImpl implements GraphListener {
 				.setNumber(String.valueOf(parentRun.getNumber()))
 				.setStartTime(TimingAction.getStartTime(stepStartNode))
 				.setCauses(CIEventCausesFactory.processCauses(stepStartNode));
-		OctaneSDK.getInstance().getEventsService().publishEvent(event);
+		OctaneSDK.getClients().forEach(octaneClient -> octaneClient.getEventsService().publishEvent(event));
 	}
 
 	private void sendStageFinishedEvent(StepEndNode stepEndNode) {
@@ -151,7 +150,7 @@ public class WorkflowListenerOctaneImpl implements GraphListener {
 				.setDuration(TimingAction.getStartTime(stepEndNode) - TimingAction.getStartTime(stepStartNode))
 				.setResult(extractFlowNodeResult(stepEndNode))
 				.setCauses(CIEventCausesFactory.processCauses(stepEndNode));
-		OctaneSDK.getInstance().getEventsService().publishEvent(event);
+		OctaneSDK.getClients().forEach(octaneClient -> octaneClient.getEventsService().publishEvent(event));
 	}
 
 	private CIBuildResult extractFlowNodeResult(FlowNode node) {
