+++ src/main/java/com/microfocus/application/automation/tools/octane/executor/TestExecutionJobCreatorService.java
@@ -28,7 +28,7 @@ import com.hp.octane.integrations.dto.executor.TestExecutionInfo;
 import com.hp.octane.integrations.dto.executor.TestSuiteExecutionInfo;
 import com.hp.octane.integrations.dto.executor.impl.TestingToolType;
 import com.hp.octane.integrations.dto.scm.SCMRepository;
-import com.hp.octane.integrations.util.SdkConstants;
+import com.hp.octane.integrations.utils.SdkConstants;
 import com.microfocus.application.automation.tools.model.ResultsPublisherModel;
 import com.microfocus.application.automation.tools.octane.actions.UFTTestDetectionPublisher;
 import com.microfocus.application.automation.tools.octane.executor.scmmanager.ScmPluginFactory;
