+++ src/main/java/com/microfocus/application/automation/tools/octane/model/processors/projects/WorkFlowJobProcessor.java
@@ -22,10 +22,7 @@
 
 package com.microfocus.application.automation.tools.octane.model.processors.projects;
 
-import com.hp.octane.integrations.dto.DTOFactory;
-import com.hp.octane.integrations.dto.configuration.OctaneConfiguration;
 import com.microfocus.application.automation.tools.octane.configuration.ConfigurationService;
-import com.microfocus.application.automation.tools.octane.configuration.ServerConfiguration;
 import com.microfocus.application.automation.tools.octane.tests.build.BuildHandlerUtils;
 import hudson.model.Cause;
 import hudson.model.Job;
@@ -45,8 +42,6 @@ import java.util.List;
  */
 
 public class WorkFlowJobProcessor extends AbstractProjectProcessor<WorkflowJob> {
-	private static final DTOFactory dtoFactory = DTOFactory.getInstance();
-
 	WorkFlowJobProcessor(Job job) {
 		super((WorkflowJob) job);
 	}
@@ -68,21 +63,9 @@ public class WorkFlowJobProcessor extends AbstractProjectProcessor<WorkflowJob> 
 
 			//  TODO: support parameters
 		}
-		this.job.scheduleBuild(delay, new Cause.RemoteCause(getOctaneConfiguration() == null ? "non available URL" : getOctaneConfiguration().getUrl(), "octane driven execution"));
-	}
-
-	private OctaneConfiguration getOctaneConfiguration() {
-		OctaneConfiguration result = null;
-		ServerConfiguration serverConfiguration = ConfigurationService.getServerConfiguration();
-		if (serverConfiguration.location != null && !serverConfiguration.location.isEmpty() &&
-				serverConfiguration.sharedSpace != null && !serverConfiguration.sharedSpace.isEmpty()) {
-			result = dtoFactory.newDTO(OctaneConfiguration.class)
-					.setUrl(serverConfiguration.location)
-					.setSharedSpace(serverConfiguration.sharedSpace)
-					.setApiKey(serverConfiguration.username)
-					.setSecret(serverConfiguration.password.getPlainText());
-		}
-		return result;
+		this.job.scheduleBuild(delay, new Cause.RemoteCause(ConfigurationService.getServerConfiguration() == null ?
+				"non available URL" :
+				ConfigurationService.getServerConfiguration().location, "octane driven execution"));
 	}
 
 	@Override
