+++ src/main/java/com/microfocus/application/automation/tools/octane/tests/xml/TestResultXmlWriter.java
@@ -95,9 +95,9 @@ public class TestResultXmlWriter {
 			writer.writeStartElement("build");
 			writer.writeAttribute("server_id", ConfigurationService.getModel().getIdentity());
 			writer.writeAttribute("job_id", buildDescriptor.getJobId());
-			writer.writeAttribute("job_name", buildDescriptor.getJobName());
+			//writer.writeAttribute("job_name", buildDescriptor.getJobName());
 			writer.writeAttribute("build_id", buildDescriptor.getBuildId());
-			writer.writeAttribute("build_name", buildDescriptor.getBuildName());
+			//writer.writeAttribute("build_name", buildDescriptor.getBuildName());
 			if (!StringUtils.isEmpty(buildDescriptor.getSubType())) {
 				writer.writeAttribute("sub_type", buildDescriptor.getSubType());
 			}
