+++ src/main/java/com/microfocus/application/automation/tools/settings/OctaneServerSettingsBuilder.java
@@ -22,6 +22,7 @@
 
 package com.microfocus.application.automation.tools.settings;
 
+import com.hp.octane.integrations.OctaneConfiguration;
 import com.hp.octane.integrations.OctaneSDK;
 import com.microfocus.application.automation.tools.model.OctaneServerSettingsModel;
 import com.microfocus.application.automation.tools.octane.CIJenkinsServicesImpl;
@@ -52,10 +53,11 @@ import org.apache.logging.log4j.Logger;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
 
+import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.util.Collections;
-import java.util.Date;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.UUID;
 
@@ -64,242 +66,282 @@ import java.util.UUID;
  */
 
 public class OctaneServerSettingsBuilder extends Builder {
-	private static final Logger logger = LogManager.getLogger(OctaneServerSettingsBuilder.class);
+    private static final Logger logger = LogManager.getLogger(OctaneServerSettingsBuilder.class);
 
-	@Override
-	public OctaneDescriptorImpl getDescriptor() {
-		return (OctaneDescriptorImpl) super.getDescriptor();
-	}
+    @Override
+    public OctaneDescriptorImpl getDescriptor() {
+        return (OctaneDescriptorImpl) super.getDescriptor();
+    }
 
-	/**
-	 * Descriptor for {@link OctaneServerSettingsBuilder}. Used as a singleton. The class is marked as
-	 * public so that it can be accessed from views.
-	 * <p>
-	 * <p>
-	 * See <tt>src/main/resources/hudson/plugins/hello_world/HelloWorldBuilder/*.jelly</tt> for the
-	 * actual HTML fragment for the configuration screen.
-	 */
+    /**
+     * Descriptor for {@link OctaneServerSettingsBuilder}. Used as a singleton. The class is marked as
+     * public so that it can be accessed from views.
+     * <p>
+     * <p>
+     * See <tt>src/main/resources/hudson/plugins/hello_world/HelloWorldBuilder/*.jelly</tt> for the
+     * actual HTML fragment for the configuration screen.
+     */
 
-	@Extension
-	public static final class OctaneDescriptorImpl extends BuildStepDescriptor<Builder> {
+    @Extension
+    public static final class OctaneDescriptorImpl extends BuildStepDescriptor<Builder> {
 
-		@CopyOnWrite
-		private OctaneServerSettingsModel[] servers;
+        @CopyOnWrite
+        private OctaneServerSettingsModel[] servers;
 
-		@Override
-		protected XmlFile getConfigFile() {
-			XmlFile xmlFile = super.getConfigFile();
-			//Between 5.1 to 5.2 - migration hp->hpe was done.
-			//Old configuration file 'com.hp.application.automation.tools.settings.OctaneServerSettingsBuilder.xml'
-			//is replaced by new one 'com.hpe.application.automation.tools.settings.OctaneServerSettingsBuilder.xml'.
-			//As well, inside the configuration, there were replaces of hp->hpe
-			//if xmlFile is not exist, we will check if configuration file name exist in format of 5.1 version
-			//if so, we will copy old configuration to new one with replacements of hp->hpe
-			if (!xmlFile.exists()) {
-				//try to get from old path
-				File oldConfigurationFile = new File(xmlFile.getFile().getPath().replace("hpe", "hp"));
-				if (oldConfigurationFile.exists()) {
-					try {
-						String configuration = FileUtils.readFileToString(oldConfigurationFile);
-						String newConfiguration = StringUtils.replace(configuration, ".hp.", ".hpe.");
-						FileUtils.writeStringToFile(xmlFile.getFile(), newConfiguration);
-						xmlFile = super.getConfigFile();
-					} catch (IOException e) {
-						logger.error("failed to copy ALM Octane Plugin configuration 5.1 to new 5.2 format : " + e.getMessage());
-					}
-				}
-			}
+        private transient List<OctaneConfiguration> octaneConfigurations = new LinkedList<>();
 
-			return xmlFile;
-		}
+        @Override
+        protected XmlFile getConfigFile() {
+            XmlFile xmlFile = super.getConfigFile();
+            //Between 5.1 to 5.2 - migration hp->hpe was done.
+            //Old configuration file 'com.hp.application.automation.tools.settings.OctaneServerSettingsBuilder.xml'
+            //is replaced by new one 'com.hpe.application.automation.tools.settings.OctaneServerSettingsBuilder.xml'.
+            //As well, inside the configuration, there were replaces of hp->hpe
+            //if xmlFile is not exist, we will check if configuration file name exist in format of 5.1 version
+            //if so, we will copy old configuration to new one with replacements of hp->hpe
+            if (!xmlFile.exists()) {
+                //try to get from old path
+                File oldConfigurationFile = new File(xmlFile.getFile().getPath().replace("hpe", "hp"));
+                if (oldConfigurationFile.exists()) {
+                    try {
+                        String configuration = FileUtils.readFileToString(oldConfigurationFile);
+                        String newConfiguration = StringUtils.replace(configuration, ".hp.", ".hpe.");
+                        FileUtils.writeStringToFile(xmlFile.getFile(), newConfiguration);
+                        xmlFile = super.getConfigFile();
+                    } catch (IOException e) {
+                        logger.error("failed to copy ALM Octane Plugin configuration 5.1 to new 5.2 format : " + e.getMessage());
+                    }
+                }
+            }
 
-		public OctaneDescriptorImpl() {
-			load();
+            return xmlFile;
+        }
 
-			OctaneServerSettingsModel model = getModel();
-			if (StringUtils.isEmpty(model.getIdentity())) {
-				model.setIdentity(UUID.randomUUID().toString());
-				model.setIdentityFrom(new Date().getTime());
-				save();
-			}
+        public OctaneDescriptorImpl() {
+            load();
+            // todo: validate & clean || migrate old config and save
 
-			try {
-				OctaneSDK.init(new CIJenkinsServicesImpl());
-			} catch (IllegalStateException ise) {
-				logger.warn("Octane SDK already initialized");
-			}
-		}
+        }
 
-		private static ServerConfiguration convertToServerConfiguration(OctaneServerSettingsModel model) {
-			return new ServerConfiguration(
-					model.getLocation(),
-					model.getSharedSpace(),
-					model.getUsername(),
-					model.getPassword(),
-					model.getImpersonatedUser(),
-					model.isSuspend());
-		}
+        @Inject
+        public void postInitialize() {
+            if (!isInitialConfigValid(servers)) {
+                return;
+            }
+            for (OctaneServerSettingsModel innerServerConfiguration : servers) {
+                OctaneConfiguration octaneConfiguration = new OctaneConfiguration(innerServerConfiguration.getIdentity(), innerServerConfiguration.getLocation(),
+                        innerServerConfiguration.getSharedSpace());
+                octaneConfiguration.setClient(innerServerConfiguration.getUsername());
+                octaneConfiguration.setSecret(innerServerConfiguration.getPassword().getPlainText());
+                octaneConfigurations.add(octaneConfiguration);
+                OctaneSDK.addClient(octaneConfiguration, CIJenkinsServicesImpl.class);
+            }
+        }
 
-		@Override
-		public boolean isApplicable(
-				@SuppressWarnings("rawtypes") Class<? extends AbstractProject> aClass) {
-			// Indicates that this builder can be used with all kinds of project
-			// types
-			return true;
-		}
+        private boolean isInitialConfigValid(OctaneServerSettingsModel[] servers) {
+            if (servers == null) {
+                return false;
+            } else if (servers.length > 1) {
+                return true;
+            }
+            OctaneServerSettingsModel innerServerConfiguration = servers[0];
+            if (innerServerConfiguration.getLocation() == null || innerServerConfiguration.getLocation().isEmpty()) {
+                return false;
+            }
+            return true;
+        }
 
-		/**
-		 * This human readable name is used in the configuration screen.
-		 */
-		@Override
-		public String getDisplayName() {
-			return "";
-		}
+        private static ServerConfiguration convertToServerConfiguration(OctaneServerSettingsModel model) {
+            return new ServerConfiguration(
+                    model.getLocation(),
+                    model.getSharedSpace(),
+                    model.getUsername(),
+                    model.getPassword(),
+                    model.getImpersonatedUser(),
+                    model.isSuspend());
+        }
 
-		@Override
-		public boolean configure(StaplerRequest req, JSONObject formData) throws FormException {
+        @Override
+        public boolean isApplicable(
+                @SuppressWarnings("rawtypes") Class<? extends AbstractProject> aClass) {
+            // Indicates that this builder can be used with all kinds of project
+            // types
+            return true;
+        }
 
-			JSONObject jsonObject = (JSONObject) formData.get("mqm");
-			List<OctaneServerSettingsModel> list = req.bindJSONToList(OctaneServerSettingsModel.class, jsonObject);
-			OctaneServerSettingsModel newModel = list.get(0);
+        /**
+         * This human readable name is used in the configuration screen.
+         */
+        @Override
+        public String getDisplayName() {
+            return "";
+        }
 
-			if (jsonObject.containsKey("showIdentity")) {
-				JSONObject showIdentityJo = (JSONObject) jsonObject.get("showIdentity");
-				String identity = showIdentityJo.getString("identity");
-				validateConfiguration(doCheckInstanceId(identity), "Plugin instance id");
+        @Override
+        public boolean configure(StaplerRequest req, JSONObject formData) throws FormException {
+            JSONObject jsonObject = (JSONObject) formData.get("mqm");
+            List<OctaneServerSettingsModel> list = req.bindJSONToList(OctaneServerSettingsModel.class, jsonObject);
+            //	todo: list above contains full list of all configurations
+            //	todo: traverse the list and
+            //		1) apply changes on known instance ID configurations
+            //		2) remove missing configurations
+            //		3) add new configurations
+            OctaneServerSettingsModel newModel = list.get(0);
 
-				OctaneServerSettingsModel oldModel = getModel();
-				if (!oldModel.getIdentity().equals(identity)) {
-					newModel.setIdentity(identity);
-				}
-			}
-			setModel(newModel);
-			return super.configure(req, formData);
-		}
+            if (jsonObject.containsKey("showIdentity")) {
+                JSONObject showIdentityJo = (JSONObject) jsonObject.get("showIdentity");
+                String identity = showIdentityJo.getString("identity");
+                validateConfiguration(doCheckInstanceId(identity), "Plugin instance id");
 
-		public void setModel(OctaneServerSettingsModel newModel) {
-			//infer uiLocation
-			MqmProject mqmProject;
-			try {
-				mqmProject = ConfigurationParser.parseUiLocation(newModel.getUiLocation());
-				newModel.setSharedSpace(mqmProject.getSharedSpace());
-				newModel.setLocation(mqmProject.getLocation());
-			} catch (FormValidation fv) {
-				logger.warn("tested configuration failed on Octane URL parse: " + fv.getMessage(), fv);
-			}
+                OctaneServerSettingsModel oldModel = getModel();
+                if (!oldModel.getIdentity().equals(identity)) {
+                    newModel.setIdentity(identity);
+                }
+            }
+            setModel(newModel);
+            return super.configure(req, formData);
+        }
 
-			OctaneServerSettingsModel oldModel = getModel();
+        public void setModel(OctaneServerSettingsModel newModel) {
+            //infer uiLocation
+            MqmProject mqmProject;
+            try {
+                mqmProject = ConfigurationParser.parseUiLocation(newModel.getUiLocation());
+                newModel.setSharedSpace(mqmProject.getSharedSpace());
+                newModel.setLocation(mqmProject.getLocation());
+            } catch (FormValidation fv) {
+                logger.warn("tested configuration failed on Octane URL parse: " + fv.getMessage(), fv);
+            }
 
-			//set identity in new model
-			boolean identityChanged = StringUtils.isNotEmpty(newModel.getIdentity());
-			if (!identityChanged) { //set identity from old model
-				newModel.setIdentity(oldModel.getIdentity());
-				newModel.setIdentityFrom(oldModel.getIdentityFrom());
-			}
+            OctaneServerSettingsModel oldModel = getModel();
 
-			servers = (new OctaneServerSettingsModel[]{newModel});
-			save();
+            //  set identity in new model
+            if (newModel.getIdentity() != null && !newModel.getIdentity().isEmpty()) {
+                System.out.println("keeping the new identity from UI: " + newModel.getIdentity());
+            } else if (oldModel.getIdentity() != null && !oldModel.getIdentity().isEmpty()) {
+                newModel.setIdentity(oldModel.getIdentity());
+                newModel.setIdentityFrom(oldModel.getIdentityFrom());
+            } else {
+                newModel.setIdentity(UUID.randomUUID().toString());
+                newModel.setIdentityFrom(System.currentTimeMillis());
+            }
 
-			ServerConfiguration oldConfiguration = convertToServerConfiguration(oldModel);
-			ServerConfiguration newConfiguration = convertToServerConfiguration(newModel);
-			if (!oldConfiguration.equals(newConfiguration)) {
-				fireOnChanged(newConfiguration, oldConfiguration);
-			}
-		}
+            servers = (new OctaneServerSettingsModel[]{newModel});
+            OctaneConfiguration octaneConfiguration = octaneConfigurations.stream()
+                    .filter(oc -> oc.getInstanceId().equals(newModel.getIdentity()))
+                    .findFirst()
+                    .orElse(new OctaneConfiguration(newModel.getIdentity(), newModel.getLocation(), newModel.getSharedSpace()));
 
-		private void fireOnChanged(ServerConfiguration configuration, ServerConfiguration oldConfiguration) {
-			OctaneSDK.getInstance().getConfigurationService().notifyChange();
-			ExtensionList<ConfigurationListener> listeners = ExtensionList.lookup(ConfigurationListener.class);
-			for (ConfigurationListener listener : listeners) {
-				try {
-					listener.onChanged(configuration, oldConfiguration);
-				} catch (ThreadDeath t) {
-					throw t;
-				} catch (Throwable t) {
-					logger.warn(t);
-				}
-			}
-		}
+            octaneConfiguration.setSharedSpace(newModel.getSharedSpace());
+            octaneConfiguration.setUrl(newModel.getLocation());
+            octaneConfiguration.setClient(newModel.getUsername());
+            octaneConfiguration.setSecret(newModel.getPassword().getPlainText());
+            if (!octaneConfigurations.contains(octaneConfiguration)) {
+                octaneConfigurations.add(octaneConfiguration);
+                OctaneSDK.addClient(octaneConfiguration, CIJenkinsServicesImpl.class);
+            }
+            save();
 
-		@SuppressWarnings("unused")
-		public FormValidation doTestConnection(@QueryParameter("uiLocation") String uiLocation,
-		                                       @QueryParameter("username") String username,
-		                                       @QueryParameter("password") String password,
-		                                       @QueryParameter("impersonatedUser") String impersonatedUser) {
-			MqmProject mqmProject;
-			try {
-				mqmProject = ConfigurationParser.parseUiLocation(uiLocation);
-			} catch (FormValidation fv) {
-				logger.warn("tested configuration failed on Octane URL parse: " + fv.getMessage(), fv);
-				return fv;
-			}
+            ServerConfiguration oldConfiguration = convertToServerConfiguration(oldModel);
+            ServerConfiguration newConfiguration = convertToServerConfiguration(newModel);
+            if (!oldConfiguration.equals(newConfiguration)) {
+                fireOnChanged(newConfiguration, oldConfiguration);
+            }
+        }
 
-			//  if parse is good, check authentication/authorization
-			ConfigurationParser parser = Jenkins.getInstance().getExtensionList(ConfigurationParser.class).iterator().next();
-			FormValidation validation = parser.checkConfiguration(mqmProject.getLocation(), mqmProject.getSharedSpace(), username, Secret.fromString(password));
+        private void fireOnChanged(ServerConfiguration configuration, ServerConfiguration oldConfiguration) {
+            OctaneSDK.getClients().forEach(octaneClient -> octaneClient.getConfigurationService().notifyChange());
+            ExtensionList<ConfigurationListener> listeners = ExtensionList.lookup(ConfigurationListener.class);
+            for (ConfigurationListener listener : listeners) {
+                try {
+                    listener.onChanged(configuration, oldConfiguration);
+                } catch (ThreadDeath t) {
+                    throw t;
+                } catch (Throwable t) {
+                    logger.warn(t);
+                }
+            }
+        }
 
-			//  if still good, check Jenkins user permissions
-			try {
-				SecurityContext preserveContext = impersonate(impersonatedUser);
-				if (!Jenkins.getInstance().hasPermission(Item.READ)) {
-					logger.warn("tested configuration failed on insufficient Jenkins' user permissions");
-					validation = FormValidation.errorWithMarkup(ConfigurationParser.markup("red", Messages.JenkinsUserPermissionsFailure()));
-				}
-				depersonate(preserveContext);
-			} catch (FormValidation fv) {
-				logger.warn("tested configuration failed on impersonating Jenkins' user, most likely non existent user provided", fv);
-				return fv;
-			}
+        @SuppressWarnings("unused")
+        public FormValidation doTestConnection(@QueryParameter("uiLocation") String uiLocation,
+                                               @QueryParameter("username") String username,
+                                               @QueryParameter("password") String password,
+                                               @QueryParameter("impersonatedUser") String impersonatedUser) {
+            MqmProject mqmProject;
+            try {
+                mqmProject = ConfigurationParser.parseUiLocation(uiLocation);
+            } catch (FormValidation fv) {
+                logger.warn("tested configuration failed on Octane URL parse: " + fv.getMessage(), fv);
+                return fv;
+            }
 
-			return validation;
-		}
+            //  if parse is good, check authentication/authorization
+            ConfigurationParser parser = Jenkins.getInstance().getExtensionList(ConfigurationParser.class).iterator().next();
+            FormValidation validation = parser.checkConfiguration(mqmProject.getLocation(), mqmProject.getSharedSpace(), username, Secret.fromString(password));
 
-		public OctaneServerSettingsModel[] getServers() {
-			if (servers == null) {
-				servers = new OctaneServerSettingsModel[]{new OctaneServerSettingsModel()};
-			}
-			return servers;
-		}
+            //  if still good, check Jenkins user permissions
+            try {
+                SecurityContext preserveContext = impersonate(impersonatedUser);
+                if (!Jenkins.getInstance().hasPermission(Item.READ)) {
+                    logger.warn("tested configuration failed on insufficient Jenkins' user permissions");
+                    validation = FormValidation.errorWithMarkup(ConfigurationParser.markup("red", Messages.JenkinsUserPermissionsFailure()));
+                }
+                depersonate(preserveContext);
+            } catch (FormValidation fv) {
+                logger.warn("tested configuration failed on impersonating Jenkins' user, most likely non existent user provided", fv);
+                return fv;
+            }
 
-		public OctaneServerSettingsModel getModel() {
-			return getServers()[0];
-		}
+            return validation;
+        }
 
-		public ServerConfiguration getServerConfiguration() {
-			return convertToServerConfiguration(getModel());
-		}
+        public OctaneServerSettingsModel[] getServers() {
+            if (servers == null) {
+                servers = new OctaneServerSettingsModel[]{new OctaneServerSettingsModel()};
+            }
+            return servers;
+        }
 
-		private SecurityContext impersonate(String user) throws FormValidation {
-			SecurityContext originalContext = null;
-			if (user != null && !user.equalsIgnoreCase("")) {
-				User jenkinsUser = User.get(user, false, Collections.emptyMap());
-				if (jenkinsUser != null) {
-					originalContext = ACL.impersonate(jenkinsUser.impersonate());
-				} else {
-					throw FormValidation.errorWithMarkup(ConfigurationParser.markup("red", Messages.JenkinsUserPermissionsFailure()));
-				}
-			}
-			return originalContext;
-		}
+        public OctaneServerSettingsModel getModel() {
+            return getServers()[0];
+        }
 
-		private void depersonate(SecurityContext originalContext) {
-			if (originalContext != null) {
-				ACL.impersonate(originalContext.getAuthentication());
-			}
-		}
+        public ServerConfiguration getServerConfiguration() {
+            return convertToServerConfiguration(getModel());
+        }
 
-		public FormValidation doCheckInstanceId(@QueryParameter String value) {
-			if (StringUtils.isBlank(value)) {
-				return FormValidation.error("Plugin Instance Id cannot be empty");
-			}
+        private SecurityContext impersonate(String user) throws FormValidation {
+            SecurityContext originalContext = null;
+            if (user != null && !user.equalsIgnoreCase("")) {
+                User jenkinsUser = User.get(user, false, Collections.emptyMap());
+                if (jenkinsUser != null) {
+                    originalContext = ACL.impersonate(jenkinsUser.impersonate());
+                } else {
+                    throw FormValidation.errorWithMarkup(ConfigurationParser.markup("red", Messages.JenkinsUserPermissionsFailure()));
+                }
+            }
+            return originalContext;
+        }
 
-			return FormValidation.ok();
-		}
+        private void depersonate(SecurityContext originalContext) {
+            if (originalContext != null) {
+                ACL.impersonate(originalContext.getAuthentication());
+            }
+        }
 
-		private void validateConfiguration(FormValidation result, String formField) throws FormException {
-			if (!result.equals(FormValidation.ok())) {
-				throw new FormException("Validation of property in ALM Octane Server Configuration failed: " + result.getMessage(), formField);
-			}
-		}
-	}
+        public FormValidation doCheckInstanceId(@QueryParameter String value) {
+            if (StringUtils.isBlank(value)) {
+                return FormValidation.error("Plugin Instance Id cannot be empty");
+            }
+
+            return FormValidation.ok();
+        }
+
+        private void validateConfiguration(FormValidation result, String formField) throws FormException {
+            if (!result.equals(FormValidation.ok())) {
+                throw new FormException("Validation of property in ALM Octane Server Configuration failed: " + result.getMessage(), formField);
+            }
+        }
+    }
 }
