+++ src/test/java/com/microfocus/application/automation/tools/octane/configuration/ConfigApiTest.java
@@ -26,32 +26,19 @@ import com.gargoylesoftware.htmlunit.FailingHttpStatusCodeException;
 import com.gargoylesoftware.htmlunit.HttpMethod;
 import com.gargoylesoftware.htmlunit.Page;
 import com.gargoylesoftware.htmlunit.WebRequest;
-import com.gargoylesoftware.htmlunit.html.HtmlForm;
-import com.gargoylesoftware.htmlunit.html.HtmlPage;
+import com.microfocus.application.automation.tools.octane.PlugInAbstractTest;
 import hudson.util.Secret;
 import net.jcip.annotations.NotThreadSafe;
 import net.sf.json.JSONObject;
-import org.junit.*;
-import org.jvnet.hudson.test.JenkinsRule;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.UUID;
 
 // import com.gargoylesoftware.htmlunit.WebRequestSettings;
 @SuppressWarnings({"squid:S2699","squid:S3658","squid:S2259","squid:S1872","squid:S2925","squid:S109","squid:S1607","squid:S2701","squid:S2698"})
 @NotThreadSafe
-public class ConfigApiTest {
-
-	@ClassRule
-	static final public JenkinsRule rule = new JenkinsRule();
-	static final private JenkinsRule.WebClient client = rule.createWebClient();
-
-	@Before
-	public void init() throws Exception {
-		HtmlPage configPage = client.goTo("configure");
-		HtmlForm form = configPage.getFormByName("config");
-		form.getInputByName("_.uiLocation").setValueAttribute("http://localhost:8008/ui/?p=1001/1002");
-		form.getInputByName("_.username").setValueAttribute("username");
-		form.getInputByName("_.password").setValueAttribute("password");
-		rule.submit(form);
-	}
+public class ConfigApiTest extends PlugInAbstractTest {
 
 	@Test
 	public void testRead() throws Exception {
@@ -59,7 +46,7 @@ public class ConfigApiTest {
 		String configAsString = page.getWebResponse().getContentAsString();
 		JSONObject config = JSONObject.fromObject(configAsString);
 		Assert.assertEquals("http://localhost:8008", config.getString("location"));
-		Assert.assertEquals("1001", config.getString("sharedSpace"));
+		Assert.assertEquals(PlugInAbstractTest.ssp, config.getString("sharedSpace"));
 		Assert.assertEquals("username", config.getString("username"));
 		Assert.assertEquals(ConfigurationService.getModel().getIdentity(), config.getString("serverIdentity"));
 	}
@@ -72,48 +59,51 @@ public class ConfigApiTest {
 		// basic scenario: location, shared space and credentials
 		JSONObject config = new JSONObject();
 		config.put("location", "http://localhost:8088");
-		config.put("sharedSpace", "1001");
+		String sharedSP = UUID.randomUUID().toString();
+		config.put("sharedSpace", sharedSP);
 		config.put("username", "username1");
 		config.put("password", "password1");
-		JenkinsRule.WebClient wc = rule.createWebClient();
-		WebRequest req = new WebRequest(wc.createCrumbedUrl("nga/configuration/save"), HttpMethod.POST);
+		WebRequest req = new WebRequest(client.createCrumbedUrl("nga/configuration/save"), HttpMethod.POST);
 		req.setEncodingType(null);
 		req.setRequestBody(config.toString());
-		Page page = wc.getPage(req);
+		Page page = client.getPage(req);
 		config = JSONObject.fromObject(page.getWebResponse().getContentAsString());
-		checkConfig(config, "http://localhost:8088", "1001", "username1", Secret.fromString("password1"));
+		checkConfig(config, "http://localhost:8088", sharedSP, "username1", Secret.fromString("password1"));
 		Assert.assertEquals(ConfigurationService.getModel().getIdentity(), config.getString("serverIdentity"));
 
 
 		// location, shared space, no credentials
 		config = new JSONObject();
 		config.put("location", "http://localhost:8888");
-		config.put("sharedSpace", "1002");
+		String sharedSP1 = UUID.randomUUID().toString();
+		config.put("sharedSpace", sharedSP1);
 		req.setRequestBody(config.toString());
 		page = client.getPage(req);
 		config = JSONObject.fromObject(page.getWebResponse().getContentAsString());
-		checkConfig(config, "http://localhost:8888", "1002", "username1", Secret.fromString("password1"));
+		checkConfig(config, "http://localhost:8888", sharedSP1, "username1", Secret.fromString("password1"));
 		Assert.assertEquals(ConfigurationService.getModel().getIdentity(), config.getString("serverIdentity"));
 //
 		// location, shared space and username without password
 		config = new JSONObject();
 		config.put("location", "http://localhost:8882");
-		config.put("sharedSpace", "1003");
+		String sharedSP2 = UUID.randomUUID().toString();
+		config.put("sharedSpace", sharedSP2);
 		config.put("username", "username3");
 		req.setRequestBody(config.toString());
 		page = client.getPage(req);
 		config = JSONObject.fromObject(page.getWebResponse().getContentAsString());
-		checkConfig(config, "http://localhost:8882", "1003", "username3", Secret.fromString(""));
+		checkConfig(config, "http://localhost:8882", sharedSP2, "username3", Secret.fromString(""));
 		Assert.assertEquals(ConfigurationService.getModel().getIdentity(), config.getString("serverIdentity"));
 
 		// uiLocation and identity
 		config = new JSONObject();
-		config.put("uiLocation", "http://localhost:8881/ui?p=1001/1002");
+		String sharedSP3 = UUID.randomUUID().toString();
+		config.put("uiLocation", "http://localhost:8881/ui?p=" + sharedSP3 + "/1002");
 		config.put("serverIdentity", "2d2fa955-1d13-4d8c-947f-ab11c72bf850");
 		req.setRequestBody(config.toString());
 		page = client.getPage(req);
 		config = JSONObject.fromObject(page.getWebResponse().getContentAsString());
-		checkConfig(config, "http://localhost:8881", "1001", "username3", Secret.fromString(""));
+		checkConfig(config, "http://localhost:8881", sharedSP3, "username3", Secret.fromString(""));
 		Assert.assertEquals("2d2fa955-1d13-4d8c-947f-ab11c72bf850", config.getString("serverIdentity"));
 		Assert.assertEquals("2d2fa955-1d13-4d8c-947f-ab11c72bf850", ConfigurationService.getModel().getIdentity());
