+++ src/test/java/com/microfocus/application/automation/tools/octane/configuration/ConfigurationListenerTest.java
@@ -24,38 +24,29 @@ package com.microfocus.application.automation.tools.octane.configuration;
 
 import com.gargoylesoftware.htmlunit.html.HtmlForm;
 import com.gargoylesoftware.htmlunit.html.HtmlPage;
+import com.microfocus.application.automation.tools.octane.PlugInAbstractTest;
 import com.microfocus.application.automation.tools.octane.tests.ExtensionUtil;
-import org.junit.*;
-import org.jvnet.hudson.test.JenkinsRule;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
 import org.jvnet.hudson.test.TestExtension;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ConfigurationListenerTest {
-
-	@ClassRule
-	static final public JenkinsRule rule = new JenkinsRule();
+public class ConfigurationListenerTest extends PlugInAbstractTest {
 
 	private TestConfigurationListener listener;
 
 	@Before
-	public void init() throws Exception {
-		HtmlPage configPage = rule.createWebClient().goTo("configure");
-		HtmlForm form = configPage.getFormByName("config");
-
-		form.getInputByName("_.uiLocation").setValueAttribute("http://localhost:8028/ui/?p=1001/1002");
-		form.getInputByName("_.username").setValueAttribute("username");
-		form.getInputByName("_.password").setValueAttribute("password");
-		rule.submit(form);
-
+	public void initialize() {
 		listener = ExtensionUtil.getInstance(rule, TestConfigurationListener.class);
 		Assert.assertEquals("Listener count doesn't match 1",1, listener.getCount());
 
 		List<ServerConfiguration> confs = listener.getConfigurationsChange();
 		Assert.assertNotNull("Configuration is null",confs);
 		Assert.assertEquals("Config size doesn't match 2",2, confs.size());
-		Assert.assertEquals("location doesn't match localhost:8020","http://localhost:8028", confs.get(0).location);
+		Assert.assertEquals("location doesn't match localhost:8008","http://localhost:8008", confs.get(0).location);
 		Assert.assertEquals("username doesn't match username","username", confs.get(0).username);
 		Assert.assertNull(confs.get(1).location);
 		Assert.assertNull(confs.get(1).username);
@@ -63,7 +54,7 @@ public class ConfigurationListenerTest {
 
 	@Test
 	public void testConfigurationListener() throws Exception {
-		HtmlPage configPage = rule.createWebClient().goTo("configure");
+		HtmlPage configPage = client.goTo("configure");
 		HtmlForm form = configPage.getFormByName("config");
 		// password is cleared upon form retrieval (restore the value)
 		form.getInputByName("_.password").setValueAttribute("password");
@@ -72,7 +63,7 @@ public class ConfigurationListenerTest {
 		rule.submit(form);
 		Assert.assertEquals("Listener count doesn't match 1",1, listener.getCount());
 
-		configPage = rule.createWebClient().goTo("configure");
+		configPage = client.goTo("configure");
 		form = configPage.getFormByName("config");
 		// increased when configuration changes
 		form.getInputByName("_.username").setValueAttribute("username2");
@@ -82,9 +73,9 @@ public class ConfigurationListenerTest {
 		List<ServerConfiguration> confs = listener.getConfigurationsChange();
 		Assert.assertNotNull("Configuration is null",confs);
 		Assert.assertEquals("Config count doesn't match 2",2, confs.size());
-		Assert.assertEquals("http://localhost:8028", confs.get(0).location);
+		Assert.assertEquals("http://localhost:8008", confs.get(0).location);
 		Assert.assertEquals("username2", confs.get(0).username);
-		Assert.assertEquals("http://localhost:8028", confs.get(1).location);
+		Assert.assertEquals("http://localhost:8008", confs.get(1).location);
 		Assert.assertEquals("username", confs.get(1).username);
 	}
