+++ src/test/java/com/microfocus/application/automation/tools/octane/tests/detection/ResultFieldsDetectionTest.java
@@ -22,6 +22,7 @@
 
 package com.microfocus.application.automation.tools.octane.tests.detection;
 
+import com.microfocus.application.automation.tools.octane.PlugInAbstractTest;
 import com.microfocus.application.automation.tools.octane.tests.CopyResourceSCM;
 import com.microfocus.application.automation.tools.octane.tests.ExtensionUtil;
 import com.microfocus.application.automation.tools.octane.tests.TestUtils;
@@ -30,46 +31,42 @@ import hudson.model.AbstractBuild;
 import hudson.model.FreeStyleProject;
 import hudson.tasks.Maven;
 import hudson.tasks.junit.JUnitResultArchiver;
-import org.junit.*;
-import org.jvnet.hudson.test.JenkinsRule;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
 import org.jvnet.hudson.test.ToolInstallations;
 import org.mockito.Mockito;
 
 import java.io.File;
 import java.io.FileReader;
+import java.util.UUID;
 
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-public class ResultFieldsDetectionTest {
-
-	@Rule
-	public final JenkinsRule rule = new JenkinsRule();
+public class ResultFieldsDetectionTest  extends PlugInAbstractTest {
 
 	private static FreeStyleProject project;
 
 	private static ResultFieldsDetectionService detectionService;
 
-	private final JenkinsRule.WebClient jClient = rule.createWebClient();
-
 	@Before
 	public void setUp() throws Exception {
 
-		project = rule.createFreeStyleProject("junit - job");
+		project = rule.createFreeStyleProject("junit - job-" + UUID.randomUUID().toString());
 		JUnitExtension junitExtension = ExtensionUtil.getInstance(rule, JUnitExtension.class);
 		detectionService = Mockito.mock(ResultFieldsDetectionService.class);
 		junitExtension._setResultFieldsDetectionService(detectionService);
 
 		Maven.MavenInstallation mavenInstallation = ToolInstallations.configureMaven3();
-		//Maven.MavenInstallation mavenInstallation = new Maven.MavenInstallation("default-system-maven", System.getenv("MAVEN_HOME"), JenkinsRule.NO_PROPERTIES);
+		//Maven.MavenInstallation mavenInstallation = new Maven.MavenInstallation("default-system-maven", TestUtils.getMavenHome(), JenkinsRule.NO_PROPERTIES);
 
-		//project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" -U test",System.getenv("MAVEN_HOME")), mavenInstallation.getName(), null, null, "-Dmaven.test.failure.ignore=true"));
+		//project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" -U test",TestUtils.getMavenHome()), mavenInstallation.getName(), null, null, "-Dmaven.test.failure.ignore=true"));
 		project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" -U test -Dmaven.repo.local=%s\\m2-temp",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")), mavenInstallation.getName(), null, null, "-Dmaven.test.failure.ignore=true"));
+				TestUtils.getMavenHome(),System.getenv("TEMP")), mavenInstallation.getName(), null, null, "-Dmaven.test.failure.ignore=true"));
 
 		project.setScm(new CopyResourceSCM("/helloWorldRoot"));
-
-		TestUtils.createDummyConfiguration();
 	}
 
 	@Test
@@ -111,8 +108,6 @@ public class ResultFieldsDetectionTest {
 		File mqmTestsXml = new File(build.getRootDir(), "mqmTests.xml");
 		ResultFieldsXmlReader xmlReader = new ResultFieldsXmlReader(new FileReader(mqmTestsXml));
 		ResultFields resultFields = xmlReader.readXml().getResultFields();
-		;
-
 		Assert.assertNull(resultFields.getFramework());
 		Assert.assertNull(resultFields.getTestingTool());
 		Assert.assertNull(resultFields.getTestLevel());
