+++ src/test/java/com/microfocus/application/automation/tools/octane/tests/detection/TestNGExtensionTest.java
@@ -23,6 +23,7 @@
 package com.microfocus.application.automation.tools.octane.tests.detection;
 
 import com.google.inject.Inject;
+import com.microfocus.application.automation.tools.octane.PlugInAbstractTest;
 import com.microfocus.application.automation.tools.octane.tests.CopyResourceSCM;
 import com.microfocus.application.automation.tools.octane.tests.TestUtils;
 import hudson.FilePath;
@@ -40,7 +41,6 @@ import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.jvnet.hudson.test.JenkinsRule;
 import org.jvnet.hudson.test.ToolInstallations;
 
 import javax.xml.stream.XMLStreamException;
@@ -52,10 +52,7 @@ import java.util.Map;
 import java.util.UUID;
 
 @SuppressWarnings({"squid:S2698","squid:S2699","squid:S3658"})
-public class TestNGExtensionTest {
-
-	@ClassRule
-	public static final JenkinsRule rule = new JenkinsRule();
+public class TestNGExtensionTest extends PlugInAbstractTest {
 
 	@ClassRule
 	public static final TemporaryFolder temporaryFolder = new TemporaryFolder();
@@ -68,7 +65,6 @@ public class TestNGExtensionTest {
 	@Before
 	public void setUp() throws Exception {
 		mavenName = ToolInstallations.configureMaven3().getName();
-		TestUtils.createDummyConfiguration();
 	}
 
 	@Test
@@ -78,7 +74,7 @@ public class TestNGExtensionTest {
 		FreeStyleProject project = rule.createFreeStyleProject(projectName);
 		project.setScm(new CopyResourceSCM("/helloWorldTestNGRoot"));
 		project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" test -Dmaven.repo.local=%s\\m2-temp",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
+				TestUtils.getMavenHome(),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
 		project.getPublishersList().add(new JUnitResultArchiver("helloWorld/target/surefire-reports/TEST*.xml, helloWorld2/target/surefire-reports/TEST*.xml"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(project);
 
@@ -92,7 +88,7 @@ public class TestNGExtensionTest {
 		FreeStyleProject project = rule.createFreeStyleProject(projectName);
 		project.setScm(new CopyResourceSCM("/helloWorldTestNGRoot/helloWorld"));
 		project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" test -Dmaven.repo.local=%s\\m2-temp",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
+				TestUtils.getMavenHome(),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
 		project.getPublishersList().add(new JUnitResultArchiver("target/surefire-reports/TEST*.xml"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(project);
 
@@ -106,7 +102,7 @@ public class TestNGExtensionTest {
 		FreeStyleProject project = rule.createFreeStyleProject(projectName);
 		project.setScm(new CopyResourceSCM("/helloWorldTestNGRoot"));
 		project.getBuildersList().add(new Maven(String.format("--settings \"%s\\conf\\settings.xml\" test -P custom-report-location -Dmaven.repo.local=%s\\m2-temp",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
+				TestUtils.getMavenHome(),System.getenv("TEMP")), mavenName, null, null, "-Dmaven.test.failure.ignore=true"));
 		project.getPublishersList().add(new JUnitResultArchiver("**\\custom-report-location/**.xml"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(project);
 
@@ -121,7 +117,7 @@ public class TestNGExtensionTest {
 		mavenProject.runHeadless();
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -Dmaven.test.failure.ignore=true");
-		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldTestNGRoot/helloWorld"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(mavenProject);
 
@@ -137,7 +133,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -Dmaven.test.failure.ignore=true");
 		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+				TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldTestNGRoot"));
 		MavenModuleSetBuild build = (MavenModuleSetBuild) TestUtils.runAndCheckBuild(mavenProject);
 
@@ -163,7 +159,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -P custom-report-location -Dmaven.test.failure.ignore=true");
 		mavenProject.setGoals(String.format("test -P custom-report-location --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+				TestUtils.getMavenHome(),System.getenv("TEMP")));
 
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldTestNGRoot/helloWorld"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(mavenProject);
@@ -183,7 +179,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -P custom-report-location -Dmaven.test.failure.ignore=true");
 		mavenProject.setGoals(String.format("test -P custom-report-location --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+				TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldTestNGRoot"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(mavenProject);
 
@@ -202,7 +198,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -P custom-report-location -Dmaven.test.failure.ignore=true");
 		mavenProject.setGoals(String.format("test -P custom-report-location --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+				TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.getPublishersList().add(new JUnitResultArchiver("**/custom-report-location/**.xml"));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldTestNGRoot"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(mavenProject);
@@ -219,7 +215,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("verify");
 		mavenProject.setGoals(String.format("verify --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",
-				System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+				TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldFailsafe"));
 		AbstractBuild build = TestUtils.runAndCheckBuild(mavenProject);
 
@@ -270,7 +266,7 @@ public class TestNGExtensionTest {
 		mavenProject.setMaven(mavenName);
 
 		//mavenProject.setGoals("test -Dmaven.test.failure.ignore=true");
-		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldRoot/helloWorld"));
 		MavenModuleSetBuild build = (MavenModuleSetBuild) TestUtils.runAndCheckBuild(mavenProject);
 
@@ -301,7 +297,7 @@ public class TestNGExtensionTest {
 		mavenProject.runHeadless();
 		mavenProject.setMaven(mavenName);
 		//mavenProject.setGoals("test -Dmaven.test.failure.ignore=true");
-		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",System.getenv("MAVEN_HOME"),System.getenv("TEMP")));
+		mavenProject.setGoals(String.format("test --settings \"%s\\conf\\settings.xml\" -Dmaven.repo.local=\"%s\\m2-temp\" -Dmaven.test.failure.ignore=true",TestUtils.getMavenHome(),System.getenv("TEMP")));
 		mavenProject.setScm(new CopyResourceSCM("/helloWorldRoot/helloWorld"));
 		MavenModuleSetBuild build = (MavenModuleSetBuild) TestUtils.runAndCheckBuild(mavenProject);
