+++ src/test/resources/helloCucumberWorld/src/test/java/MyStepdefs.java
@@ -1,5 +1,7 @@
-import cucumber.api.PendingException;
-import cucumber.api.java.en.*;
+import cucumber.api.java.en.And;
+import cucumber.api.java.en.Given;
+import cucumber.api.java.en.Then;
+import cucumber.api.java.en.When;
 
 /**
  * Created by franksha on 27/09/2016.
