+++ src/test/resources/helloCucumberWorld/src/test/java/RunnerTest.java
@@ -1,3 +1,4 @@
+import com.hpe.alm.octane.OctaneCucumber;
 import cucumber.api.CucumberOptions;
 import org.junit.runner.RunWith;
